pulsola-track-service/pyproject.toml
[project]
name = "pulsola-track-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.0.0",
    "aiofiles>=24.1.0",
    "anyio>=4.9.0",
    "fastapi>=0.115.14",
    "faststream[otel,rabbit]>=0.5.43",
    "granian>=2.5.0",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-instrumentation-tortoiseorm>=0.55b1",
    "opentelemetry-sdk>=1.34.1",
    "orjson>=3.10.18",
    "pillow==11.3.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "starlette>=0.46.2",
    "tortoise-orm[accel,asyncpg]>=0.25.1",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.2",
    "types-aioboto3-lite[s3]>=15.1.0",
    "types-aiobotocore-lite[s3]>=2.24.1",
    "types-aiofiles>=24.1.0.20250809",
]
prod = [
    "granian[uvloop]>=2.5.0",
    "uvloop>=0.21.0"
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
extend-select = ["I"]

[tool.mypy]
plugins = ['pydantic.mypy']
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["aioboto3.*"]
follow_untyped_imports = true
